requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# ── Numerical / Computer Vision ────────────────────────────────────────────────
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ── Object Detection & Tracking (YOLOv8) ──────────────────────────────────────
ultralytics>=8.2.0

# ── Face Recognition (InsightFace / ArcFace) ──────────────────────────────────
insightface>=0.7.3
onnxruntime>=1.16.0              # swap for onnxruntime-gpu for CUDA support

# ── Face Recognition Fallback ──────────────────────────────────────────────────
deepface>=0.0.90

# ── Scientific Computing ───────────────────────────────────────────────────────
scipy>=1.11.0

# ── Optional: Better Liveness Detection (MediaPipe) ───────────────────────────
mediapipe>=0.10.0
